requirements.txt
torch~=1.11.0
numpy
pandas
tensorboard
--extra-index-url https://_json_key_base64:ewogICJ0eXBlIjogInNlcnZpY2VfYWNjb3VudCIsCiAgInByb2plY3RfaWQiOiAiZ3JpZC1iYWNrZW5kLTI2NjcyMSIsCiAgInByaXZhdGVfa2V5X2lkIjogImVjYjA2MTFiODY4YTY3NDhjODFkMWEwNmZlNTEwZjdjZWQ2MjNkNDciLAogICJwcml2YXRlX2tleSI6ICItLS0tLUJFR0lOIFBSSVZBVEUgS0VZLS0tLS1cbk1JSUV2Z0lCQURBTkJna3Foa2lHOXcwQkFRRUZBQVNDQktnd2dnU2tBZ0VBQW9JQkFRQ293NERXd1IvY1pzTEZcbjRKR1Z3bHhjV1pJLzRQT1VyVWNPUVRUWTE2c3BsbzZDZ0tVa253aktoV25aVUE3RkhvSkU5a054djJtUDdPbmpcbis0bkUraUFvZWlRajhrS3RTc1BWRTZPcEF1Ym5hM0FqL3F5SStGMFJBN1NzYXZBeXMza0svYzEva3ArS09TMllcbk16YSt2UlpXVFRHTVh0NkVTeU8rN05zSDkzeFkwbmZtUm9YMDI3amVDZlNaUDdVblNMUHY0dDB3QnFpMnBNbkVcbjlCb01DNFE0SU5kYzZUWUNTeHUzS2pFNGRKRkRHb1FiQnhHTmN2SDdRYk9TRFRoOXZiU1NKeVdtNEh3QndBakJcbmJjcnVxUGJjQUN4dUh3Sm4zSWhDMHdrYkFuZjF2dTVyNm0rNFBSV0QvV2lNK2REMkpvVjNTVkpTRy9lMFlGNXdcbnp1eDMvR0NqQWdNQkFBRUNnZ0VBSjg5Q1g2WmI0UzVlYUxZaU1ZMFJJM20vblpEdmRKVnhhd1BudHZVYys2ajNcbndnSWRzcWRQT0JMRGxzOGpSTTEvRmt4dk9YQlpNdW5FZkpLVCs2S3pIa2s5cURzWURtL1NCVHZtUWRLYzdGODBcbksxR0NtcWJYc1ZGSjk5Z2NCQ0hBL2w1RGNRSUIzMlhBZ3lRbG1GWGJaSTArRUdvNm5FTnJVYmptckJqdHZMZTZcblpLRUNPSDZMQWRQLzZmaEdQNUhGYkEyczl4N3BkRXB2WXYyU3g5MDQvNW9oSE1XZ0hoRXdneThiSDVUZzNyRW5cbkl2bEEwQTErK1E0QTVETm5Md0EvV0FwTVFnNWtPV1hrbVZ3cDVnN3g1b0R1dDY0WFlNYklVTFpBNGhEN1hMMUFcbjRZZjZkbnRYY01zd0N1cUtNUWpxMzAzNWdNOWVjbm5ab2lOcHhOT0xsUUtCZ1FEcTArR3dKa21VOWhyQUtaUmRcblFySUNnKzlOM2pTb0N3b2NqYVZDWW9mZ3k0LzdIejVGc1JXL1ZOYmkzcjRQTVNtREYrZ0JrODAwd3RFdHMvZFRcblFuMkV1VmordEh0NXdQZ3JmSG95eUxiZzZxU2JyRHpPTnEzYk85cEdyWGEzQ2xWUHVLdytOb1pIZ0k5YlE0Z1lcblRXcWpaMENxMVZBS2g4dk1lbEl4SzZOblJ3S0JnUUMzK3NnazlmWVh2c2kvK1hjWVJqajNqYnNnQnZ0dTJFekRcbkFFeGUzdVMwSEU4Q0lXTjhuUS9jWFpBSVZZNTF3Y3NtNmlCeTJzUnVlczA5YUpNK1k0cDNkaVo3dWFCZ3ZIWCtcbnBoNGpzTXBlWWZzY2tRVWVUNkhsQi8wdDVLS1MyUUZBbU1BUGdLeVFaTVd0dk1tUEVhY0hMVVdlOENqbEdxQUNcbnlsMHJ4QVJoeFFLQmdRQ1RGWDJpRUlrOEpWNGlDS2ttTlBVOGdCanV4QVJsVU96WGI1MFlWSkRaSzRlV2VqNU5cbjlwb3hpbGxDSnRRU3ZlOUxiamppYkFvM1J1TXdaaTIrMThHNDFVTTUzaUFNNWVKTGtwOFRtZ1o2SUY1VUozQzZcbmxTTVdVNG5uaFJUN2x4eFNYOUI2OHpudUpVY2xtY252dHVYWlRYNEN3ck5zdFJ2Z2lxbGFwUU9uVndLQmdBL25cbk5KL0ZSeFY4WDF3QTAyT2N4bFhYd1lJZU9HTzNmTU1xWm0rWThzb0MzRzJCdDFqUk5zckVwNnVnd2FTNk1MWlhcbmJLQ2crblBXVjZGa1JiNFIycWxDNUVhem9BSmNxQkp3QjVEc09rSDRWRDErNTZOY0hORndaRmt3TjBGY1VyaGpcbnU4NWJRQTluTnBNekMzVTdnR1dsYXl3MjlFSUJrQzFOa3NveXNuSkZBb0dCQU9JbFVSLzNYWWhEU2JMbnY3Rjhcbk5PS2lLMzU4c2loQ1VmTlFIVmR2ZWdBSzFtYzQ2NWx3Z3BTYzhGL1NtY2RVT3l3RDBRUzRLczh2ZUVtbGtiWHRcbmhBUU5PeVJzQVBHbjRaQUc3RjNFVklvSmd2WUdwL2RqSlgwQWZnTFEzQjgvcFc2OGV1SmZQelNrMFJSODZ5bi9cbmVrSnc0YWJyRE5tUFE0QkR0ZjdQSGVTWlxuLS0tLS1FTkQgUFJJVkFURSBLRVktLS0tLVxuIiwKICAiY2xpZW50X2VtYWlsIjogImxpZ2h0bmluZy1weXBpLXJlcG9zaXRvcnktcmVhZEBncmlkLWJhY2tlbmQtMjY2NzIxLmlhbS5nc2VydmljZWFjY291bnQuY29tIiwKICAiY2xpZW50X2lkIjogIjExNzkxMjY1OTM1OTEwNjE0MjkwNCIsCiAgImF1dGhfdXJpIjogImh0dHBzOi8vYWNjb3VudHMuZ29vZ2xlLmNvbS9vL29hdXRoMi9hdXRoIiwKICAidG9rZW5fdXJpIjogImh0dHBzOi8vb2F1dGgyLmdvb2dsZWFwaXMuY29tL3Rva2VuIiwKICAiYXV0aF9wcm92aWRlcl94NTA5X2NlcnRfdXJsIjogImh0dHBzOi8vd3d3Lmdvb2dsZWFwaXMuY29tL29hdXRoMi92MS9jZXJ0cyIsCiAgImNsaWVudF94NTA5X2NlcnRfdXJsIjogImh0dHBzOi8vd3d3Lmdvb2dsZWFwaXMuY29tL3JvYm90L3YxL21ldGFkYXRhL3g1MDkvbGlnaHRuaW5nLXB5cGktcmVwb3NpdG9yeS1yZWFkJTQwZ3JpZC1iYWNrZW5kLTI2NjcyMS5pYW0uZ3NlcnZpY2VhY2NvdW50LmNvbSIKfQo=@us-central1-python.pkg.dev/grid-backend-266721/xpi/simple
lightning~=0.0.46
google-cloud-pubsub
requests~=2.27.1